MAJ 2021-08-08

+++ src/main/java/com/safetynet/model/Persons.java
@@ -26,7 +26,7 @@ public class Persons {
 
     private String adress;
     private String city;
-    private Long zip;
+    private Integer zip;
     private String phone;
     private String email;
 }
